+++ slsim/Deflectors/DeflectorPopulation/cluster_deflectors.py
@@ -2,13 +2,13 @@ import numpy as np
 import numpy.random as random
 from slsim.Util import param_util
 from slsim.selection import object_cut
-from slsim.Deflectors.richness2mass import mass_richness_relation
-from slsim.Deflectors.halo_population import gene_e_ang_halo, concent_m_w_scatter
+from slsim.Deflectors.MassLightConnection.richness2mass import mass_richness_relation
+from slsim.Halos.halo_population import gene_e_ang_halo, concent_m_w_scatter
 from colossus.cosmology import cosmology as colossus_cosmo
-from slsim.Deflectors.velocity_dispersion import vel_disp_abundance_matching
-from slsim.Deflectors.elliptical_lens_galaxies import elliptical_projected_eccentricity
-from slsim.Deflectors.velocity_dispersion import vel_disp_nfw
-from slsim.Deflectors.deflectors_base import DeflectorsBase
+from slsim.Deflectors.MassLightConnection.velocity_dispersion import vel_disp_abundance_matching
+from slsim.Deflectors.DeflectorPopulation.elliptical_lens_galaxies import elliptical_projected_eccentricity
+from slsim.Deflectors.MassLightConnection.velocity_dispersion import vel_disp_nfw
+from slsim.Deflectors.DeflectorPopulation.deflectors_base import DeflectorsBase
 from slsim.Deflectors.deflector import Deflector
 from lenstronomy.Util.param_util import phi_q2_ellipticity
 from astropy import units as u
