+++ slsim/Deflectors/DeflectorPopulation/compound_lens_halos_galaxies.py
@@ -1,8 +1,8 @@
 import numpy as np
 import numpy.random as random
 from slsim.selection import object_cut
-from slsim.Deflectors.deflectors_base import DeflectorsBase
-from slsim.Deflectors.elliptical_lens_galaxies import elliptical_projected_eccentricity
+from slsim.Deflectors.DeflectorPopulation.deflectors_base import DeflectorsBase
+from slsim.Deflectors.DeflectorPopulation.elliptical_lens_galaxies import elliptical_projected_eccentricity
 from slsim.Deflectors.deflector import Deflector
