+++ slsim/Deflectors/DeflectorPopulation/deflectors_base.py
