+++ slsim/Deflectors/DeflectorPopulation/elliptical_lens_galaxies.py
@@ -2,8 +2,8 @@ import numpy as np
 import numpy.random as random
 from slsim.selection import object_cut
 from slsim.Util import param_util
-from slsim.Deflectors.deflectors_base import DeflectorsBase
-from slsim.Deflectors.velocity_dispersion import vel_disp_abundance_matching
+from slsim.Deflectors.DeflectorPopulation.deflectors_base import DeflectorsBase
+from slsim.Deflectors.MassLightConnection.velocity_dispersion import vel_disp_abundance_matching
 from slsim.Deflectors.deflector import Deflector
