+++ slsim/Deflectors/DeflectorTypes/epl_sersic.py
@@ -1,6 +1,6 @@
 from slsim.Deflectors.DeflectorTypes.deflector_base import DeflectorBase
 from slsim.Util.param_util import ellipticity_slsim_to_lenstronomy
-from slsim.Deflectors.velocity_dispersion import theta_E_from_vel_disp_epl
+from slsim.Deflectors.MassLightConnection.velocity_dispersion import theta_E_from_vel_disp_epl
 
 
 class EPLSersic(DeflectorBase):
@@ -8,8 +8,8 @@ class EPLSersic(DeflectorBase):
 
     required quantities in dictionary:
     - 'vel_disp': SIS equivalent velocity dispersion of the deflector
-    - 'e1_mass': eccentricity of NFW profile
-    - 'e2_mass': eccentricity of NFW profile
+    - 'e1_mass': eccentricity of EPL profile
+    - 'e2_mass': eccentricity of EPL profile
     - 'stellar_mass': stellar mass in physical M_sol
     - 'angular_size': half-light radius of stellar/light profile in radian
     - 'n_sersic': Sersic index of deflector light
