+++ slsim/Deflectors/DeflectorTypes/nfw_cluster.py
@@ -1,5 +1,5 @@
 from slsim.Deflectors.DeflectorTypes.deflector_base import DeflectorBase
-from slsim.Deflectors.velocity_dispersion import vel_disp_nfw
+from slsim.Deflectors.MassLightConnection.velocity_dispersion import vel_disp_nfw
 from slsim.Deflectors.DeflectorTypes.epl_sersic import EPLSersic
 from slsim.Util.param_util import ellipticity_slsim_to_lenstronomy
 import numpy as np
