+++ slsim/Deflectors/MassLightConnection/light2mass.py
