+++ slsim/Deflectors/MassLightConnection/richness2mass.py
