+++ slsim/Deflectors/MassLightConnection/velocity_dispersion.py
