+++ slsim/Deflectors/__init__.py
@@ -1,9 +1,9 @@
-from .all_lens_galaxies import AllLensGalaxies
-from .compound_lens_halos_galaxies import CompoundLensHalosGalaxies
-from .elliptical_lens_galaxies import EllipticalLensGalaxies
-from .cluster_deflectors import ClusterDeflectors
+from slsim.Deflectors.DeflectorPopulation.all_lens_galaxies import AllLensGalaxies
+from slsim.Deflectors.DeflectorPopulation.compound_lens_halos_galaxies import CompoundLensHalosGalaxies
+from slsim.Deflectors.DeflectorPopulation.elliptical_lens_galaxies import EllipticalLensGalaxies
+from slsim.Deflectors.DeflectorPopulation.cluster_deflectors import ClusterDeflectors
 from .deflector import Deflector
-from .deflectors_base import DeflectorsBase
+from slsim.Deflectors.DeflectorPopulation.deflectors_base import DeflectorsBase
 
 __all__ = [
     "AllLensGalaxies",
